+++ test/e2e/chaos/podchaos/misc.go
@@ -1,0 +1,22 @@
+package podchaos
+
+import (
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/util/wait"
+	"k8s.io/client-go/kubernetes"
+	"time"
+)
+
+func waitPodRunning(name, namespace string, cli kubernetes.Interface) error {
+	return wait.Poll(5*time.Second, 5*time.Minute, func() (done bool, err error) {
+		pod, err := cli.CoreV1().Pods(namespace).Get(name, metav1.GetOptions{})
+		if err != nil {
+			return false, nil
+		}
+		if pod.Status.Phase != corev1.PodRunning {
+			return false, nil
+		}
+		return true, nil
+	})
+}
