+++ test/e2e/chaos/podchaos/pod_failure.go
@@ -1,0 +1,227 @@
+package podchaos
+
+import (
+	"context"
+	"github.com/chaos-mesh/chaos-mesh/api/v1alpha1"
+	"github.com/chaos-mesh/chaos-mesh/test/e2e/e2econst"
+	"github.com/chaos-mesh/chaos-mesh/test/e2e/util"
+	"github.com/chaos-mesh/chaos-mesh/test/pkg/fixture"
+	. "github.com/onsi/ginkgo"
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/labels"
+	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/apimachinery/pkg/util/wait"
+	"k8s.io/client-go/kubernetes"
+	"k8s.io/kubernetes/test/e2e/framework"
+	"k8s.io/utils/pointer"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"time"
+)
+
+func TestcasePodFailureOnceThenDelete(ns string, kubeCli kubernetes.Interface, cli client.Client) {
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	By("preparing experiment pods")
+	nd := fixture.NewTimerDeployment("timer", ns)
+	_, err := kubeCli.AppsV1().Deployments(ns).Create(nd)
+	framework.ExpectNoError(err, "create timer deployment error")
+	err = util.WaitDeploymentReady("timer", ns, kubeCli)
+	framework.ExpectNoError(err, "wait timer deployment ready error")
+
+	By("create pod failure chaos CRD objects")
+	listOption := metav1.ListOptions{
+		LabelSelector: labels.SelectorFromSet(map[string]string{
+			"app": "timer",
+		}).String(),
+	}
+	podFailureChaos := &v1alpha1.PodChaos{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "timer-failure",
+			Namespace: ns,
+		},
+		Spec: v1alpha1.PodChaosSpec{
+			Selector: v1alpha1.SelectorSpec{
+				Namespaces: []string{
+					ns,
+				},
+				LabelSelectors: map[string]string{
+					"app": "timer",
+				},
+			},
+			Action: v1alpha1.PodFailureAction,
+			Mode:   v1alpha1.OnePodMode,
+		},
+	}
+
+	err = cli.Create(ctx, podFailureChaos)
+	framework.ExpectNoError(err, "create pod failure chaos error")
+
+	By("waiting for assertion some pod fall into failure")
+	err = wait.PollImmediate(5*time.Second, 5*time.Minute, func() (done bool, err error) {
+		pods, err := kubeCli.CoreV1().Pods(ns).List(listOption)
+		if err != nil {
+			return false, nil
+		}
+		if len(pods.Items) != 1 {
+			return false, nil
+		}
+		pod := pods.Items[0]
+		for _, c := range pod.Spec.Containers {
+			if c.Image == e2econst.PauseImage {
+				return true, nil
+			}
+		}
+		return false, nil
+	})
+	framework.ExpectNoError(err, "failed to verify PodFailure")
+
+	By("delete pod failure chaos CRD objects")
+	err = cli.Delete(ctx, podFailureChaos)
+	framework.ExpectNoError(err, "failed to delete pod failure chaos")
+
+	By("waiting for assertion recovering")
+	err = wait.PollImmediate(5*time.Second, 1*time.Minute, func() (done bool, err error) {
+		pods, err := kubeCli.CoreV1().Pods(ns).List(listOption)
+		if err != nil {
+			return false, nil
+		}
+		if len(pods.Items) != 1 {
+			return false, nil
+		}
+		pod := pods.Items[0]
+		for _, c := range pod.Spec.Containers {
+			if c.Image == nd.Spec.Template.Spec.Containers[0].Image {
+				return true, nil
+			}
+		}
+		return false, nil
+	})
+	framework.ExpectNoError(err, "pod failure recover failed")
+}
+
+func TestcasePodFailurePauseThenUnPause(ns string, kubeCli kubernetes.Interface, cli client.Client) {
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	By("preparing experiment pods")
+	nd := fixture.NewTimerDeployment("timer", ns)
+	_, err := kubeCli.AppsV1().Deployments(ns).Create(nd)
+	framework.ExpectNoError(err, "create timer deployment error")
+	err = util.WaitDeploymentReady("timer", ns, kubeCli)
+	framework.ExpectNoError(err, "wait timer deployment ready error")
+
+	By("create pod failure chaos CRD objects")
+	var pods *corev1.PodList
+	listOption := metav1.ListOptions{
+		LabelSelector: labels.SelectorFromSet(map[string]string{
+			"app": "timer",
+		}).String(),
+	}
+
+	podFailureChaos := &v1alpha1.PodChaos{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "timer-failure",
+			Namespace: ns,
+		},
+		Spec: v1alpha1.PodChaosSpec{
+			Selector: v1alpha1.SelectorSpec{
+				Namespaces:     []string{ns},
+				LabelSelectors: map[string]string{"app": "timer"},
+			},
+			Action:   v1alpha1.PodFailureAction,
+			Mode:     v1alpha1.OnePodMode,
+			Duration: pointer.StringPtr("9m"),
+			Scheduler: &v1alpha1.SchedulerSpec{
+				Cron: "@every 10m",
+			},
+		},
+	}
+	err = cli.Create(ctx, podFailureChaos)
+	framework.ExpectNoError(err, "create pod failure chaos error")
+	chaosKey := types.NamespacedName{
+		Namespace: ns,
+		Name:      "timer-failure",
+	}
+
+	By("waiting for assertion some pod fall into failure")
+	// check whether the pod failure chaos succeeded or not
+	err = wait.PollImmediate(5*time.Second, 5*time.Minute, func() (done bool, err error) {
+		pods, err := kubeCli.CoreV1().Pods(ns).List(listOption)
+		if err != nil {
+			return false, nil
+		}
+		pod := pods.Items[0]
+		for _, c := range pod.Spec.Containers {
+			if c.Image == e2econst.PauseImage {
+				return true, nil
+			}
+		}
+		return false, nil
+	})
+
+	// pause experiment
+	By("pause pod failure chaos")
+	err = util.PauseChaos(ctx, cli, podFailureChaos)
+	framework.ExpectNoError(err, "pause chaos error")
+
+	By("waiting for assertion about chaos experiment paused")
+	err = wait.Poll(5*time.Second, 5*time.Minute, func() (done bool, err error) {
+		chaos := &v1alpha1.PodChaos{}
+		err = cli.Get(ctx, chaosKey, chaos)
+		framework.ExpectNoError(err, "get pod chaos error")
+		if chaos.Status.Experiment.Phase == v1alpha1.ExperimentPhasePaused {
+			return true, nil
+		}
+		return false, err
+	})
+	framework.ExpectNoError(err, "check paused chaos failed")
+
+	By("wait for 1 minutes and no pod failure")
+	pods, err = kubeCli.CoreV1().Pods(ns).List(listOption)
+	framework.ExpectNoError(err, "get timer pod error")
+	err = wait.Poll(5*time.Second, 1*time.Minute, func() (done bool, err error) {
+		pods, err = kubeCli.CoreV1().Pods(ns).List(listOption)
+		framework.ExpectNoError(err, "get timer pod error")
+		pod := pods.Items[0]
+		for _, c := range pod.Spec.Containers {
+			if c.Image == e2econst.PauseImage {
+				return true, nil
+			}
+		}
+		return false, nil
+	})
+	framework.ExpectError(err, "wait no pod failure failed")
+	framework.ExpectEqual(err.Error(), wait.ErrWaitTimeout.Error())
+
+	By("resume paused chaos experiment")
+	err = util.UnPauseChaos(ctx, cli, podFailureChaos)
+	framework.ExpectNoError(err, "resume chaos error")
+
+	By("waiting for assertion about pod failure happens again")
+	err = wait.Poll(5*time.Second, 5*time.Minute, func() (done bool, err error) {
+		chaos := &v1alpha1.PodChaos{}
+		err = cli.Get(ctx, chaosKey, chaos)
+		framework.ExpectNoError(err, "get pod chaos error")
+		if chaos.Status.Experiment.Phase == v1alpha1.ExperimentPhaseRunning {
+			return true, nil
+		}
+		return false, err
+	})
+	framework.ExpectNoError(err, "check resumed chaos failed")
+
+	By("waiting for assert pod failure happens again")
+	err = wait.Poll(5*time.Second, 5*time.Minute, func() (done bool, err error) {
+		pods, err = kubeCli.CoreV1().Pods(ns).List(listOption)
+		framework.ExpectNoError(err, "get timer pod error")
+		pod := pods.Items[0]
+		for _, c := range pod.Spec.Containers {
+			if c.Image == e2econst.PauseImage {
+				return true, nil
+			}
+		}
+		return false, nil
+	})
+	framework.ExpectNoError(err, "wait pod failure failed")
+}
