+++ test/e2e/e2econst/const.go
@@ -1,0 +1,7 @@
+package e2econst
+
+const (
+	PauseImage             = "gcr.io/google-containers/pause:latest"
+	ChaosMeshNamespace     = "chaos-testing"
+	ChaosControllerManager = "chaos-controller-manager"
+)
